backend/Machine Learning/Tour Recommendation/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud
google-cloud-aiplatform